requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
mssql-django==1.1.3
psycopg2-binary==2.9.3
pyodbc==4.0.32
pytz==2022.1
sqlparse==0.4.2
ipython=8.3.0
